Fixed bugs relating to the new keys relating to comma movement
Mainly revolving cases of the caret being in the start or the end of the file, such that the caret is outside the most outside comma there is. In such cases, the jump just doesn't respond.

Now, the caret will respond and will loop to the next or previous comma occasion