Added new comma-related Actions:
Ctrl + alt + comma -> Moves the caret to the right of the right comma and selects the text that is inside 2 commas. upon multiple uses, switches to the next element instead.

ctrl + alt + shift + comma -> Moves caret to the left of the left comma and selects the text that is inside 2 commas. upon multiple uses, switches to the previous element instead.

~NOTICE:
This is not at a good state with detection of which text it should actually select - currently it tries to select all/most text [non-whitespace] between 2 commas, when the desired outcome should be to select the text that is surrounded by parenthesis of a method call (since this action should typically be used for selecting an argument of a method call)