Added new actions for moving towards the right and left of the caret.
ctrl shift alt RIGHT -> Jump the caret to the start of the word that is ahead of the caret by an X amount.

ctrl shift alt LEFT -> Jump the caret to the start of the word that is behind of the caret by an X amount.

shift ctrl alt X -> Jump the caret by a specified character amount that has been preset via the Set And Jump Horizontally action.

shift ctrl X/G -> Jump each caret by a specified character amount in the menu dialog. Amount is configurable via the dialog screen

shift alt X -> Invert the direction of where you jump horizontally when you use the action Jump Horizontally (for Char jump only)