Added new actions for navigation between code blocks.
Both actions are for traversing through code blocks of the SAME nested level. (i.e., can't go from a code block of a class into a code block of a function that is nested in a class.)

Go to Next Block [shift ctrl alt CLOSE_BRACKET] -> Moves caret to the first non-whitespace character of the next block of code [ {} ].

Go to Previous Block [shift ctrl alt OPEN_BRACKET] -> Moves caret to the first non-whitespace character of the previous block of code [ {} ]

------------------------------------

Also added a small Status Bar Widget that displays the main caret's character offset. (how many characters the caret is away from the start of the file)