Added more actions for navigation between code blocks and an action that selects code that is in a code block.
Go to Inner Block [shift alt CLOSE_BRACKET] -> Moves caret to the start of the content of an inner code block if possible [ {} ]

Go to Outer Block [shift alt OPEN_BRACKET] -> Moves caret to the start of the content of an outer code block if possible [ {} ]

Select Code Block [shift ctrl alt W] -> Select the block of code the caret is in. This refers to what is inside curly braces, and curly braces only.